1st pass, use temp animationHelpers to isolate animation code
